Support deduct and release preclaimed value and refactor lifecycle
Move lifecycle related calls to a lifecycle mod. Also added
description for how the lifecycle works.
